requirements.txt
google.generativeai
pandas
numpy
streamlit
python-dotenv